Gateway now relies less on local developer API
This is a first implementation of a new way to get access to subdevices
on a gateway. It uses an alternative way to get a device list and
initial properties.

The local developer API is still used but only in passive mode where it
receives broadcasts of things such as clicks and changes to properties
of sensors.

86sw1 and 86sw2 are not really supported in this version as their type
id is currently unknown.